Inline status endpoint into server.js
