+++ controllers/robots/index.js
@@ -1,6 +1,5 @@
 'use strict';
 const express = require('express');
-const moment = require('moment');
 const sitemap = require('sitemap');
 const domains = require('config').get('domains');
 const { includes, concat } = require('lodash');
@@ -9,26 +8,8 @@ const { getBaseUrl, getAbsoluteUrl } = require('../../common/urls');
 const { getCanonicalRoutes } = require('../../common/route-helpers');
 const { legacyPagePaths, legacyFilesPath } = require('../../common/archived');
 const { noCache, sMaxAge } = require('../../middleware/cached');
-const appData = require('../../common/appData');
 
 const router = express.Router();
-
-const LAUNCH_DATE = moment();
-
-router.get('/status', (req, res) => {
-    res.setHeader('Access-Control-Allow-Origin', '*');
-    res.setHeader('Access-Control-Allow-Headers', 'Origin, X-Requested-With, Content-Type, Accept');
-    res.setHeader('Cache-Control', 'no-store,no-cache,max-age=0');
-
-    res.json({
-        APP_ENV: appData.environment,
-        DEPLOY_ID: appData.deployId,
-        COMMIT_ID: appData.commitId,
-        BUILD_NUMBER: appData.buildNumber,
-        START_DATE: LAUNCH_DATE.format('dddd, MMMM Do YYYY, h:mm:ss a'),
-        UPTIME: LAUNCH_DATE.toNow(true)
-    });
-});
 
 router.get('/robots.txt', noCache, (req, res) => {
     const isIndexable = includes(domains.indexable, req.get('host')) === true;
@@ -43,7 +24,11 @@ router.get('/robots.txt', noCache, (req, res) => {
     const text = [
         `user-agent: *`,
         `sitemap: ${getAbsoluteUrl(req, '/sitemap.xml')}`,
-        `${isIndexable ? disallowList.map(line => `disallow: ${line}`).join('\n') : 'disallow: /'}`
+        `${
+            isIndexable
+                ? disallowList.map(line => `disallow: ${line}`).join('\n')
+                : 'disallow: /'
+        }`
     ].join('\n');
     res.setHeader('Content-Type', 'text/plain');
     res.send(text);
